Move get_scroll_position to view module
The utility function is only used in the view module, so move it to
where it is used. Also add tests for the utility function and fix some
edge cases that would cause a panic.
