+++ src/scroll/mod.rs
@@ -1,2 +1,1 @@
 pub(crate) mod scroll_position;
-pub(crate) mod utils;
